fix(gms): Fixes delete references for single relationship aspects (#7211)
Deleting Container references were broken. This PR fixes that.

Root cause is #5248 which changed the MCP change type to DELETE for cases where an aspect after getting the reference removed was no longer valid. We do not support DELETE operations for MCPs.

The solution is then to execute EntityService#deleteAspect rather than emit a sync MCP.

A regression test was made to prevent this from happening again.

Co-authored-by: John Joyce <john@acryl.io>